rename Database Command
